move schema overview into separate file
